pyproject.toml
[project]
dependencies = [
    "fastapi>=0.115.12",
    "python-jose>=3.4.0",
    "gunicorn>=23.0.0",
    "uvicorn>=0.34.2",
    "pydantic>=2.11.4",
    "pandas>=2.2.3",
    "osmnx>=2.0.3",
    "networkx>=3.4.2",
    "geopandas>=1.0.1",
    "ortools>=9.12.4544",
    "folium>=0.19.6",
    "scikit-learn>=1.6.1",
    "matplotlib>=3.10.3",
    "numpy>=2.2.5",
    "geojson>=3.2.0",
    "pyclustering>=0.10.1.2",
]
description = "Delivetter Simulations API"
license = {text = "MIT"}
name = "delivetter-api"
requires-python = ">=3.10"
version = "0.1.0"
[tool]
[tool.pdm]
[tool.pdm.scripts]
dev = {cmd = "python -m scripts.run_local_server", env = {ENV = "dev"}}
format = {shell = "black . && isort ."}
lint = {shell = "pylint src"}
local_server = "uvicorn src.app:app --reload --port 8080"
prod = {cmd = "python -m scripts.run_local_server", env = {ENV = "prod"}}
server = "gunicorn src.app:app -b 0.0.0.0:8080 -k uvicorn.workers.UvicornWorker -w 4 -t 300 --reload"
set_precommit_hooks = "pre-commit install --install-hooks"
tunnel = "ngrok http 8080"
typecheck = {shell = "mypy ."}

[tool.pylint.basic]
good-names = ["e"]

[tool.pylint.messages_control]
disable = [
  "too-many-ancestors",
  "too-many-arguments",
  "too-many-boolean-expressions",
  "too-many-branches",
  "too-many-function-args",
  "too-many-instance-attributes",
  "too-many-lines",
  "too-many-locals",
  "too-many-nested-blocks",
  "too-many-public-methods",
  "too-many-return-statements",
  "too-many-statements",
  "too-few-public-methods",
  "duplicate-code",
  "unused-argument",
  "no-self-use",
  "fixme",
  "missing-docstring",
  "C0330",
  "C0326",
  "E1120",
  "redefined-builtin",
  "invalid-name",
  "no-self-argument",
  "broad-except",
  "raise-missing-from",
]

[tool.pylint.format]
extension-pkg-whitelist = ["pydantic"]
max-line-length = 99

[tool.black]
line-length = 99

[tool.isort]
filter_files = true
line_length = 99
profile = "black"
py_version = 310

[tool.mypy]
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
exclude = "__pypackages__"
follow_imports = "skip"
ignore_missing_imports = true
no_implicit_optional = true
plugins = [
  "pydantic.mypy",
]
python_version = "3.10"
show_column_numbers = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true

[build-system]
build-backend = "pdm.pep517.api"
requires = ["pdm-pep517>=0.12.0"]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pre-commit>=4.2.0",
    "flake8>=7.2.0",
]
